server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.isort]
profile = "black"
src_paths = ["server"]

[tool.taskipy.settings]
runner = "dotenv run"

[tool.taskipy.tasks]
pre_format = "isort ."
format = "black server"
run = "flask run"
add-migration = "flask db migrate"
apply-migrations = "flask db upgrade"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
gpiozero = "^1.5.1"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.5.3"
flask-cors = "^3.0.9"
"RPi.GPIO" = "^0.7.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
taskipy = "^1.3.0"
isort = "^5.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
